fix(parser): properly calculate start offset for sub-elements
Sub-elements for elements that span more than one line (i.e. 'target'
and 'node' for cross-reference) may appear in a line other than the
first line that the parent element spans.
An additional condition is check to properly determined the start line
of the sub-elements and thus the start row and column.